Create test script to run
